+++ src/app/app.module.ts
@@ -7,7 +7,7 @@ import { RecipesComponent } from './components/recipes/recipes.component';
 import { ShoppingListComponent } from './components/shopping/shopping-list/shopping-list.component';
 import { ListEditComponent } from './components/shopping/list-edit/list-edit.component';
 import { HeaderComponent } from './components/header/header.component';
-import { RecipeItemComponent } from './components/recipes/recipe-item/recipe-item.component';
+import { RecipeItemComponent } from './components/recipes/recipe-list/recipe-item/recipe-item.component';
 import { RecipeDetailComponent } from './components/recipes/recipe-detail/recipe-detail.component';
 import { RecipeListComponent } from "./components/recipes/recipe-list/recipe-list.component";
