--- src/app/components/recipes/recipe-item/recipe-item.component.ts
@@ -1,18 +1,0 @@
-import { Component, OnInit, Input } from '@angular/core';
-import {Recipe} from "../recipe.model";
-
-@Component({
-  selector: 'app-recipe-item',
-  templateUrl: './recipe-item.component.html',
-  styleUrls: ['./recipe-item.component.css']
-})
-export class RecipeItemComponent implements OnInit {
-  @Input() recipe!: Recipe;
-
-  constructor() {
-  }
-
-  ngOnInit(): void {
-  }
-
-}
