+++ src/app/components/recipes/recipe-list/recipe-item/recipe-item.component.html
@@ -1,6 +1,7 @@
 <a
   href="#"
-  class="list-group-item clearfix">
+  class="list-group-item clearfix"
+  (click)="onSelected()">
   <div class="float-left">
     <h4 class="list-group-item-heading">{{ recipe.name }}</h4>
     <p class="list-group-item-text">{{ recipe.description }}</p>
