+++ src/app/components/recipes/recipe-list/recipe-item/recipe-item.component.ts
@@ -1,0 +1,23 @@
+import {Component, OnInit, Input, EventEmitter, Output} from '@angular/core';
+import {Recipe} from "../../recipe.model";
+
+@Component({
+  selector: 'app-recipe-item',
+  templateUrl: './recipe-item.component.html',
+  styleUrls: ['./recipe-item.component.css']
+})
+export class RecipeItemComponent implements OnInit {
+  @Input() recipe: Recipe;
+  @Output() recipeSelected = new EventEmitter<void>();
+
+  constructor() {
+  }
+
+  ngOnInit(): void {
+  }
+
+  onSelected() {
+    this.recipeSelected.emit();
+  }
+
+}
