+++ src/app/components/recipes/recipe-list/recipe-list.component.html
@@ -9,6 +9,7 @@
     <app-recipe-item
       *ngFor="let recipeEl of recipes"
       [recipe]="recipeEl"
+      (recipeSelected)="onRecipeSelected(recipeEl)"
     ></app-recipe-item>
   </div>
 </div>
